feat: moved faqs into common file
