--- packages/shared/src/components/plus/common.ts
@@ -1,3 +1,0 @@
-export interface CommonPlusPageProps {
-  shouldShowPlusHeader?: boolean;
-}
